RiTool/SBP/backend/requirement.txt
asgiref==3.3.4
bcrypt==3.2.0
cffi==1.14.5
cryptography==3.4.7
Django==3.2.4
django-cors-headers==3.7.0
django-friendship==1.9.1
djangorestframework==3.12.4
paramiko==2.7.2
pika==1.2.0
pycparser==2.20
PyNaCl==1.4.0
pytz==2021.1
six==1.16.0
sqlparse==0.4.1
